Refactor: modularize logic into internal packages (AHH, CMS, Percentiles)
